first version of a Calabash extension that calls Daffodil
This version doesn't include scala-library.jar, as it was just bloat.
This cuts down the project size from 12MB to 4MB.

Removed jline and scallop from transitive dependencies of Daffodil,
as they're not used when calling daffodil from the API.
ICU charset and locale SPI jars are still in lib, as those don't
seem to be in maven repositories.