Updated to latest version of dependencies
Added a shell script to run calabash with
Updated to daffodil v0.14.0-SNAPSHOT and scala 2.10
Also added license text to top of each source file,
in NOTICE.txt, and in src/main/resources/META-INF so that
it automatically gets included in the jar.
Tweaked calabash.sh to be a little more robust.
Upgraded calabash to the latest released version (v1.0.18-95).